feat(OrderManagement): Add payment method formatting and status badge components
- Introduce formatPaymentMethod utility for displaying payment method labels in Vietnamese.
- Create StatusBadge component to visually represent order statuses with corresponding colors and icons.
- Refactor OrderCard and OrderDetails components to utilize new formatting and badge components for improved UI consistency.
- Implement OrderTimeline component to display the order status progression visually.
- Add OrderDetailsSkeleton for loading states in order details view.
