+++ src/lib/format-payment-method.tsx
@@ -1,0 +1,14 @@
+import { PaymentMethod } from "@/data/enum";
+
+export const formatPaymentMethod = (method: string) => {
+  switch (method) {
+    case PaymentMethod.COD:
+      return "Thanh toán khi nhận hàng";
+    case PaymentMethod.CREDIT_CARD:
+      return "Thẻ tín dụng";
+    case PaymentMethod.BANK_TRANSFER:
+      return "Chuyển khoản ngân hàng";
+    default:
+      return method;
+  }
+};
